requirements.txt
speedtest-cli
cronus
